requirements.txt
# Core requirements
streamlit==1.22.0
pandas==1.5.3
# scikit-learn==1.2.2
numpy==1.24.3
pyyaml==6.0
joblib==1.2.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# ML models
xgboost==1.7.5

# Excel handling
openpyxl==3.1.2
XlsxWriter==3.1.2

# Testing (optional)
pytest==7.4.0